requirements.txt
Django==1.8
argparse==1.2.1
diff-match-patch==20110725.1
dj-database-url==0.3.0
dj-static==0.0.6
django-import-export==0.2.9
django-toolbelt==0.0.1
gunicorn==19.4.1
psycopg2==2.6.1
static3==0.6.1
tablib==0.10.0
wsgiref==0.1.2
